azure_functions/requirements.txt
azure-functions==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
h5py==3.11.0
idna==3.8
numpy==2.1.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
stripe==11.4.1
simplejson==3.20.1